+++ interface/CPage.cpp
@@ -77,7 +77,7 @@ STDMETHODIMP CPage::Write(WORD wAddr, SAFEARRAY *psaValue)
 
 	for (int i = 0; i < UBound - LBound + 1; i++)
 	{
-		WriteByte(wAddr + i, lpData[i]);
+		WriteByte((WORD)(wAddr + i), lpData[i]);
 	}
 	SafeArrayUnaccessData(psaValue);
 	return S_OK;
